Simplify initCopyProgram for readability and a possible performance gain.
Branches can harm performance so removed the multiple branches of the if-else. Setting the copyOut to a default value is cheap so we always do that with no checking. We run one if with &&s which allows breaking out early if any of the conditions fail. If all pass we do the expensive operation (compared to setting a vec4) of sampling the texture.