Don't needlessly run the upscaling if the video is not playing.
This functionality looks to have been in the code previously, using the `loaded_count` var to decide whether to rerun the processing loop. Currently it is only updated in code that is commented out. We now use `copyVideo` for the same purpose, which is updated when the video is playing/stopped/paused.